Ignore dropped connections on a different socket
If a display id is associated with a socket that isn't the one that is
being dropped, don't disassociate. That dropped connection is stale.

This is important, but not sufficient to resolve the out of order
disconnection problem for presence. It won't resolve the issue across
pods. That's why the heartbeat record into the shared database is used.