Report folder watching correctly
`meta:[folderPath]:displays` is a SET that contains any display id that is
watching the folder. This should be used instead of folder:[folderPath]
since that is a list of files in the folder which could be empty if all
files have been removed from the folder.

`meta:[folderPath]:displays` is set during folder watch as part of
`addDisplayToMany`.