Include podname as display connection key value
This fixes issue after step 4 described in the previous commit

The pod name is stored in the connection key value, so if a
disconnection comes in after a connection, the pod value won't match and
the disconnection won't be recorded.

There is still one scenario that could lead to inconsistency.
If an ungraceful disconnection occurs client side, and then the pod
responsible for that connection goes down, the missed heartbeat won't
record the missed connection. This could be cleared up by removing the
requirement that the missed heartbeat is handled by the responsible pod.
Removing that requirement without having every pod send a disconnection
message would add significant complexity, so this scenario is ignored
for now.