Consider spark id on incoming and missed heartbeats
If an out of order redis PUBSUB missed heartbeat message arrives, the
google pubsub message won't be sent unless the spark id matches or is
missing. This way if there was a near simultaneous new connection the
disconnection won't be published.

On incoming heartbeat, publish the pubsub message if the spark id does
not match.