Don't repeat a pubsub connection on heartbeat
There are cases where a displays has two simultaneous connections on
different pods. When the heartbeat comes in and records the spark id for
one, it will overwrite the spark id for the other. This should not
result in a new pubsub conection message.

In 7764ee6d the spark check was added for missed heartbeats, which is
still needed. At the time the incoming heartbeat case was likely added
as well just for consistency. It is not needed and causes this problem
on duplicate persistent connections.